fix(T32): Reorganize pages structure
