+++ src/app/app-routing.module.ts
@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { AboutComponent } from './pages/about/about.component';
 import { AuthGuard } from './shared/guard/auth.guard';
 import { ListGuard } from './shared/guard/list.guard';
 
@@ -8,7 +7,7 @@ const routes: Routes = [
   {
     path: '',
     loadChildren: () =>
-      import('./components/shopping-list/shopping-list.module').then(
+      import('./pages/shopping-list/shopping-list.module').then(
         m => m.ShoppingListModule
       ),
     data: {
@@ -23,7 +22,7 @@ const routes: Routes = [
   {
     path: 'login',
     loadChildren: () =>
-      import('./components/sign-in/sign-in.module').then(m => m.SignInModule),
+      import('./pages/sign-in/sign-in.module').then(m => m.SignInModule),
     data: {
       title: 'Smart Shopping',
       description: 'Smart Shopping for everyone',
@@ -35,7 +34,7 @@ const routes: Routes = [
   {
     path: 'register',
     loadChildren: () =>
-      import('./components/sign-up/sign-up.module').then(m => m.SignUpModule),
+      import('./pages/sign-up/sign-up.module').then(m => m.SignUpModule),
     data: {
       title: 'Smart Shopping',
       description: 'Smart Shopping for everyone',
@@ -47,9 +46,7 @@ const routes: Routes = [
   {
     path: 'dashboard',
     loadChildren: () =>
-      import('./components/dashboard/dashboard.module').then(
-        m => m.DashboardModule
-      ),
+      import('./pages/dashboard/dashboard.module').then(m => m.DashboardModule),
     data: {
       title: 'Smart Shopping',
       description: 'Smart Shopping for everyone',
@@ -63,7 +60,7 @@ const routes: Routes = [
     path: 'forgot-password',
 
     loadChildren: () =>
-      import('./components/forgot-passoword/forgot-password.module').then(
+      import('./pages/forgot-passoword/forgot-password.module').then(
         m => m.ForgotPassowrdModule
       ),
     data: {
@@ -77,7 +74,7 @@ const routes: Routes = [
   {
     path: 'verify-email-address',
     loadChildren: () =>
-      import('./components/verify-email/verify-email.module').then(
+      import('./pages/verify-email/verify-email.module').then(
         m => m.VerifyEmailModule
       ),
     data: {
