+++ src/app/pages/forgot-passoword/forgot-password.module.ts
