+++ src/app/pages/shopping-list/done-task/done-task.component.spec.ts
@@ -10,20 +10,19 @@ import {
   TranslateModule,
 } from '@ngx-translate/core';
 import { environment } from 'src/environments/environment';
-import { FirebaseService } from '../../shared/services/firebase.service';
-import { TaskService } from '../../shared/services/task.service';
 import { provideMockStore, MockStore } from '@ngrx/store/testing';
-import { TodoTaskComponent } from './todo-task.component';
+import { DoneTaskComponent } from './done-task.component';
+import { TaskService } from 'src/app/shared/services/task.service';
 
-describe('TodoTaskComponent', () => {
-  let component: TodoTaskComponent;
-  let fixture: ComponentFixture<TodoTaskComponent>;
+describe('DoneTaskComponent', () => {
+  let component: DoneTaskComponent;
+  let fixture: ComponentFixture<DoneTaskComponent>;
   const initialState = {};
   let store: MockStore;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [TodoTaskComponent],
+      declarations: [DoneTaskComponent],
       imports: [
         provideFirebaseApp(() => initializeApp(environment.firebase)),
         provideFirestore(() => getFirestore()),
@@ -37,17 +36,13 @@ describe('TodoTaskComponent', () => {
           },
         }),
       ],
-      providers: [
-        TaskService,
-        FirebaseService,
-        provideMockStore({ initialState }),
-      ],
+      providers: [TaskService, provideMockStore({ initialState })],
     }).compileComponents();
   });
 
   beforeEach(() => {
     store = TestBed.inject(MockStore);
-    fixture = TestBed.createComponent(TodoTaskComponent);
+    fixture = TestBed.createComponent(DoneTaskComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
