+++ src/app/pages/shopping-list/shopping-list.component.spec.ts
