+++ src/app/pages/shopping-list/shopping-list.component.ts
