+++ src/app/pages/shopping-list/todo-task/todo-task.component.scss
