+++ src/app/pages/sign-up/sign-up.component.scss
