+++ src/app/pages/verify-email/verify-email.component.spec.ts
