+++ src/app/pages/verify-email/verify-email.component.ts
